Portfolio added
Portfolio is almost finished and added to the github. There is a few things I need to work on such as adding projects to the projects section of the portfolio and create a new page for each project information and button for each project live demo website and code on github.